backend/requirements.txt
boto3==1.37.33
fastapi==0.115.12
mangum==0.19.0
python-dotenv==1.1.0
uvicorn==0.34.1
